refactor: StackView class 삭제
