+++ Calculator/Calculator/Controller/CalculatorViewController.swift
@@ -120,8 +120,7 @@ class CalculatorViewController: UIViewController {
     }
     
     private func addStackView() {
-        let subStackView: StackView = StackView()
-        let content: UIStackView = subStackView.configure(operator: operatorLabel, operand: operandLabel)
+        let content: UIStackView = configure(operator: operatorLabel, operand: operandLabel)
         previousContentStackView.addArrangedSubview(content)
     }
     
@@ -134,6 +133,35 @@ class CalculatorViewController: UIViewController {
     private func clearLabel() {
         operandLabel.text = "0"
         operatorLabel.text = ""
+    }
+    
+    private func configure(operator currentOperator: UILabel, operand currentOperand: UILabel) -> UIStackView {
+        let recordedOperatorLabel: UILabel = configureItem(with: currentOperator)
+        let recordedOperandLabel: UILabel = configureItem(with: currentOperand)
+        let content: UIStackView = configureContent(item: recordedOperatorLabel, recordedOperandLabel)
+        
+        return content
+    }
+        
+    private func configureContent(item formula: UILabel...) -> UIStackView {
+        let content: UIStackView = UIStackView()
+        content.translatesAutoresizingMaskIntoConstraints = false
+        content.axis = .horizontal
+        content.alignment = .fill
+        content.distribution = .equalSpacing
+        content.spacing = 8
+        formula.forEach { content.addArrangedSubview($0) }
+        
+        return content
+    }
+
+    private func configureItem(with label: UILabel) -> UILabel {
+        let recordedLabel: UILabel = UILabel()
+        recordedLabel.font = .preferredFont(forTextStyle: .title3)
+        recordedLabel.text = label.text
+        recordedLabel.textColor = .white
+        
+        return recordedLabel
     }
 }
